requirements.txt
asgiref==3.6.0
Django==4.2.1
django-mysql==4.9.0
djangorestframework==3.14.0
dnspython==2.3.0
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.1.1
protobuf==3.20.3
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
